question-answering-system-with-T5/requirements.txt
Flask
Flask-Cors
torch
transformers
pdfplumber
ollama
groq
python-dotenv
